+++ src/bin/ode_test_rk4.rs
@@ -22,7 +22,7 @@ fn main() {
 
 fn test_fn(st: &mut State<f64>) {
     let x = st.param;
-    let y = &st.state;
+    let y = &st.value;
     let dy = &mut st.deriv;
     dy[0] = (5f64*x.powi(2) - y[0]) / (x + y[0]).exp();
 }